+++ CHANGES.md
@@ -1,3 +1,18 @@
+#### 1.5.2 / 2022-02-09
+
+##### Debug
+* added `dim` color modifier; refactoring `logGroup` method
+* changed opaque color for `^` modifier
+* delayed value creation - if the `data` is a function, then we replace it with the value returned from the function
+
+##### Other
+* added outline on `:focus` for `SelectItemControl` control
+* improved CSS on various screen widths
+
+#### 1.5.1 / 2022-01-25
+* small fixes for compatibility with WP 5.9
+* static property `$zukit_handle` of `zukit_Blocks` class is public now
+
 #### 1.5.0 / 2022-01-25
 * color modifiers support for opaque colors in `Debug` module
 * some helpers for input/textarea manipulations (compatible with React) in `JQ` module